sources/assets-gce-events/requirements.txt
google-cloud-storage==2.8.0
functions_framework==3.3.0
